Moved statepair and create_state_pair
